requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
python-multipart
requests
chromadb
sentence-transformers
numpy
torch